requirements.txt
mujoco
loguru
pyyaml
scipy
torch
onnxruntime
pynput
ipdb
termcolor
sshkeyboard
pygame
zmq
tqdm

pyvicon-datastream